+++ provider.go
@@ -14,11 +14,12 @@ func Provider() *schema.Provider {
 	return &schema.Provider{
 		ConfigureFunc: providerConfigure,
 		DataSourcesMap: map[string]*schema.Resource{
-			"clouddk_locations": dataSourceLocations(),
-			"clouddk_packages":  dataSourcePackages(),
-			"clouddk_server":    dataSourceServer(),
-			"clouddk_servers":   dataSourceServers(),
-			"clouddk_templates": dataSourceTemplates(),
+			"clouddk_locations":    dataSourceLocations(),
+			"clouddk_packages":     dataSourcePackages(),
+			"clouddk_server":       dataSourceServer(),
+			"clouddk_server_disks": dataSourceServerDisks(),
+			"clouddk_servers":      dataSourceServers(),
+			"clouddk_templates":    dataSourceTemplates(),
 		},
 		ResourcesMap: map[string]*schema.Resource{},
 		Schema: map[string]*schema.Schema{