+++ CHANGELOG.md
@@ -5,5 +5,6 @@ FEATURES:
 * **New Data Source:** `clouddk_locations`
 * **New Data Source:** `clouddk_packages`
 * **New Data Source:** `clouddk_server`
+* **New Data Source:** `clouddk_server_disks`
 * **New Data Source:** `clouddk_servers`
 * **New Data Source:** `clouddk_templates`